型を幅指定に変更

+++ FlightComponents/CommonInstrument/FlightComputer/DescentDetector.cpp
@@ -3,12 +3,12 @@
 
 /// @brief コンストラクタ
 /// @param sensitivity 大きいほど強い平滑化(0~1)
-DescentDetector::DescentDetector(double sensitivity) {
+DescentDetector::DescentDetector(float sensitivity) {
   _sensitivity = sensitivity;
 }
 /// @brief 高度を更新します。値に応じて_isDescendinも更新されます
 /// @param altitude 高度
-void DescentDetector::updateAltitude(double altitude)
+void DescentDetector::updateAltitude(float altitude)
 {
   // 指数平滑化移動平均
   _average = _averageOrigin + _sensitivity * (altitude - _averageOrigin);
